+++ mne_nirs/io/fold/__init__.py
@@ -1,0 +1,5 @@
+# Author: Robert Luke <mail@robertluke.net>
+#
+# License: BSD (3-clause)
+
+from ._fold import fold_landmark_specificity, fold_channel_specificity