+++ doc/changelog.rst
@@ -45,6 +45,8 @@ Enhancements
 * Add :func:`mne_nirs.statistics.read_glm`. By `Robert Luke`_.
 * Add :func:`mne_nirs.experimental_design.longest_inter_annotation_interval`. By `Robert Luke`_.
 * Add :func:`mne_nirs.experimental_design.drift_high_pass`. By `Robert Luke`_.
+* Add :func:`mne_nirs.io.fold_landmark_specificity`. By `Robert Luke`_.
+* Add :func:`mne_nirs.io.fold_channel_specificity`. By `Robert Luke`_.
 
 
 Fixes